Eliminate all mechanism for reserved names.
I believe that in Bawden's original syntactic closures code, this
mechanism was intended to detect mistakes such as (DEFINE A (+ A B)),
i.e. premature references to variables that would not yet have
values.  This fails if we delay the classification of the right-hand
sides of definitions, though, which we must do in order to support
forward macro references; it doesn't catch all errors anyway, since
it is a static, syntactic check; and it introduces complications that
are not worth the trouble.