Eliminate uses of EQUAL? in SYNTAX-RULES compiler.
Substitute EQ?, which is at least consistent with the rest of the
code.  I am not sure whether it is correct -- I think that it may be
better to use the macro transformer's name comparator procedure --, but
it seems to work.

EQUAL? in Chicken and Gambit would blow up when given circular
structures, and would also go to the unfortunate (and, I think,
misguided -- see Mr Pitman's thoughts on the subject) trouble of
descending into records with catastrophic effect.