Fix the classifier so that it will close keyword names in the environments from which they were obtained, and adjust the form passed to the keywords' procedures to use the name, rather than the original form, in the operator position.
This will fix an obscure problem with a macro FOO that expands to a
name BAR closed in some environment, and a use of FOO in an operator
position in an environment where BAR is not bound, if we compile the
resulting code in the context of a hygiene-aware module system.

Furthermore, classifier and transformer procedures may now reliably
assume that the operator positions of the forms they are given are
the names by which they were invoked.  (I can't imagine any reason
why they would care to know the original form.)